LabCV_OCR/requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
editdistance==0.5.3
gluoncv==0.3.0
graphviz==0.8.4
idna==2.9
imageio==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lmdb==0.98
matplotlib==3.2.1
mxnet==1.4.1
networkx==2.4
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.43.0
urllib3==1.25.8
